requirements.txt
Flask==2.1.0
pytube3=9.6.4
opencv-python==4.8.1.78
tensorflow==2.15
